requirements.txt
numpy
accelerate>=0.16.0,<1
click>=8.0.4,<9
datasets>=2.10.0,<3
deepspeed>=0.8.3,<0.9
transformers>=4.28.1,<5
tensorboardX==2.6
aim>=3.4.0
azure-storage-blob